adding search-functionality to overview-component
