+++ components/Searchbar.js
@@ -1,31 +1,4 @@
 import styled from "styled-components";
-import Fuse from "fuse.js";
-import { useState } from "react";
-import { useEffect } from "react";
-
-const fuseOptions = {
-  // isCaseSensitive: false,
-  // includeScore: false,
-  // shouldSort: true,
-  // includeMatches: false,
-  // findAllMatches: false,
-  // minMatchCharLength: 1,
-  // location: 0,
-  threshold: 0.3,
-  // distance: 100,
-  // useExtendedSearch: false,
-  // ignoreLocation: false,
-  // ignoreFieldNorm: false,
-  // fieldNormWeight: 1,
-  keys: [
-    "name",
-    "botanical_name",
-    "water_need",
-    "fertiliser_season",
-    "light_need",
-    "care_instructions",
-  ],
-};
 
 const StyledSearchBox = styled.input`
   background-color: var(--secondary-bg-color);
@@ -36,25 +9,7 @@ const StyledSearchBox = styled.input`
   font-weight: bold;
 `;
 
-export default function Searchbar({ plants }) {
-  const [results, setResults] = useState([]);
-  const [fuse, setFuse] = useState(null);
-
-  useEffect(() => {
-    setFuse(new Fuse(plants, fuseOptions));
-  }, []);
-
-  function handleSearch(event) {
-    if (!fuse) {
-      return;
-    }
-    const searchPattern = event.target.value;
-    const searchResult = fuse.search(searchPattern).slice(0, 10);
-    // .slice(0, 10) will ensure there will never be more than 10 results
-
-    /* setResults(searchResult.map((result) => result.item.name)); */
-    setResults(searchResult.map((result) => result.item));
-  }
+export default function Searchbar({ onChange }) {
   return (
     <>
       <label htmlFor="search"></label>
@@ -62,7 +17,7 @@ export default function Searchbar({ plants }) {
         type="text"
         id="search"
         placeholder="Search . . .🔎"
-        onChange={handleSearch}
+        onChange={onChange}
       />
     </>
   );
